Update README.md with rust nightly build, also the gcc version on the HLRN has changed Fix the nightly version in rust-toolchain.toml Update rand dependency split controller_it.rs into separate files so that we can use logging in the different examples
