+++ Cargo.toml
@@ -9,7 +9,7 @@ edition = "2021"
 serde = { version = "1.0.137", features = ["derive"] }
 quick-xml = { version = "0.23.0", features = ["serialize"] }
 flate2 = { version = "1.0.24" }
-rand = "0.7.3"
+rand = "0.8.5"
 metis = "0.1.1"
 flexi_logger = { version = "0.22", features = ["async"] }
 log = "0.4"
