+++ tests/three_link_it.rs
@@ -1,5 +1,6 @@
 use rust_q_sim::config::Config;
 use rust_q_sim::controller;
+use rust_q_sim::logging::init_logging;
 
 #[test]
 fn three_link_network() {
@@ -12,20 +13,10 @@ fn three_link_network() {
         .num_parts(3)
         .build();
 
+    let _logger_guard = init_logging(&config.output_dir);
+
     controller::run(config);
 
     // somehow test the output
     println!("Done");
-}
-
-#[test]
-fn equil_scenario() {
-    let config = Config::builder()
-        .network_file(String::from("./assets/equil/equil-network.xml"))
-        .population_file(String::from("./assets/equil/equil-plans.xml.gz"))
-        .output_dir(String::from("./test_output/controller_it/equil_scenario"))
-        .num_parts(2)
-        .build();
-
-    controller::run(config);
 }
